+++ src/Elmish.WPF/Binding.fs
@@ -6,6 +6,7 @@ open System.Windows.Data
 
 type Getter<'model> = 'model -> obj
 type Setter<'model,'msg> = obj -> 'model -> 'msg
+type ValidSetter<'model,'msg> = obj -> 'model -> Result<'msg,string>
 type Execute<'model,'msg> = obj -> 'model -> 'msg
 type CanExecute<'model> = obj -> 'model -> bool
 
@@ -27,6 +28,7 @@ and ViewBindings<'model,'msg> = ViewBinding<'model,'msg> list
 and Variable<'model,'msg> =
     | Bind of Getter<'model>
     | BindTwoWay of Getter<'model> * Setter<'model,'msg>
+    | BindTwoWayValidation of Getter<'model> * ValidSetter<'model,'msg>
     | BindCmd of Execute<'model,'msg> * CanExecute<'model>
     | BindModel of Getter<'model> * ViewBindings<'model,'msg>
     | BindMap of Getter<'model> * (obj -> obj)
@@ -44,6 +46,9 @@ module Binding =
             | BindTwoWay (getter,setter) -> 
                 (toModel >> getter, fun v m -> (toModel m) |> (setter v >> toMsg))
                 |> BindTwoWay
+            | BindTwoWayValidation (getter,setter) -> 
+                (toModel >> getter, fun v m -> (toModel m) |> (setter v >> Result.map toMsg))
+                |> BindTwoWayValidation
             | BindCmd (exec,canExec) ->
                 ((fun v m -> (toModel m) |> exec v |> toMsg), (fun v m -> (toModel m) |> canExec v))
                 |> BindCmd
@@ -59,15 +64,53 @@ module Binding =
 
     // Helper functions that clean up binding creation
 
-    let oneWay (getter: 'model -> 'a) p : ViewBinding<'model,'msg> = 
-        p, Bind (getter >> unbox)
-    let twoWay (getter: 'model -> 'a) (setter: 'a -> 'model -> 'msg) p : ViewBinding<'model,'msg> = 
-        p, BindTwoWay (getter >> unbox, fun v m -> setter (v :?> 'a) m)
-    let cmd exec p : ViewBinding<'model,'msg> = 
-        p, BindCmd (exec, fun _ _ -> true)
-    let cmdIf exec canExec p : ViewBinding<'model,'msg> = 
-        p, BindCmd (exec, canExec)
-    let model (getter: 'model -> '_model) (viewBinding: ViewBindings<'_model,'_msg>) (toMsg: '_msg -> 'msg) p : ViewBinding<'model,'msg> = 
-        p, BindModel (getter >> unbox, viewBinding |> mapViewBinding getter toMsg)
-    let oneWayMap (getter: 'model -> 'a) (mapper: 'a -> 'b) p : ViewBinding<'model,'msg> =
-        p, BindMap (getter >> unbox, unbox >> mapper >> unbox)
+    ///<summary>Source to target binding (i.e. BindingMode.OneWay)</summary>
+    ///<param name="getter">Gets value from the model</param>
+    ///<param name="name">Binding name</param>
+    let oneWay (getter: 'model -> 'a) name : ViewBinding<'model,'msg> = 
+        name, Bind (getter >> unbox)
+    
+    ///<summary>Either source to target or target to source (i.e. BindingMode.TwoWay)</summary>
+    ///<param name="getter">Gets value from the model</param>
+    ///<param name="setter">Setter function, returns a message to dispatch, typically to set the value in the model</param>
+    ///<param name="name">Binding name</param>
+    let twoWay (getter: 'model -> 'a) (setter: 'a -> 'model -> 'msg) name : ViewBinding<'model,'msg> = 
+        name, BindTwoWay (getter >> unbox, fun v m -> setter (v :?> 'a) m)
+    
+    ///<summary>Either source to target or target to source (i.e. BindingMode.TwoWay) with INotifyDataErrorInfo implementation)</summary>
+    ///<param name="getter">Gets value from the model</param>
+    ///<param name="setter">Validation function, returns a Result with the message to dispatch or an error string</param>
+    ///<param name="name">Binding name</param>
+    let twoWayValidation (getter: 'model -> 'a) (setter: 'a -> 'model -> Result<'msg,string>) name : ViewBinding<'model,'msg> = 
+        name, BindTwoWayValidation (getter >> unbox, fun v m -> setter (v :?> 'a) m)
+        
+    ///<summary>Command binding</summary>
+    ///<param name="exec">Execute function, returns a message to dispatch</param>
+    ///<param name="name">Binding name</param>
+    let cmd exec name : ViewBinding<'model,'msg> = 
+        name, BindCmd (exec, fun _ _ -> true)
+        
+    ///<summary>Conditional command binding</summary>
+    ///<param name="exec">Execute function, returns a message to dispatch</param>
+    ///<param name="canExec">CanExecute function, returns a bool</param>
+    ///<param name="name">Binding name</param>
+    let cmdIf exec canExec name : ViewBinding<'model,'msg> = 
+        name, BindCmd (exec, canExec)
+        
+    ///<summary>Sub-view binding</summary>
+    ///<param name="getter">Gets the sub-model from the base model</param>
+    ///<param name="viewBinding">Set of view bindings for the sub-view</param>
+    ///<param name="toMsg">Maps sub-messages to the base message type</param>
+    ///<param name="name">Binding name</param>
+    let model (getter: 'model -> '_model) (viewBinding: ViewBindings<'_model,'_msg>) (toMsg: '_msg -> 'msg) name : ViewBinding<'model,'msg> = 
+        name, BindModel (getter >> unbox, viewBinding |> mapViewBinding getter toMsg)
+        
+    ///<summary>One-way binding that applies a map when passing data to the view.
+    /// Should be used for data that a view needs wrapped in some view-specific type. 
+    /// For example when graphing a series, the data can be stored as a plain array in the model, 
+    /// and then mapped to a SeriesCollection for the view.</summary>
+    ///<param name="getter">Gets the value from the model</param>
+    ///<param name="mapper">Maps the value for consumption by the view</param>
+    ///<param name="name">Binding name</param>
+    let oneWayMap (getter: 'model -> 'a) (mapper: 'a -> 'b) name : ViewBinding<'model,'msg> =
+        name, BindMap (getter >> unbox, unbox >> mapper >> unbox)