+++ src/Samples/Elmish.Samples.DesignDataContext/Program.fs
@@ -11,7 +11,7 @@ module State =
           Age = 28
           Address = "1 Anderson Lane"
           Phone = "64212"
-          Postal = "N1N 1N1"
+          Postal = "N1N1N1"
           IsDialogVisible = false
           DialogResult = None }
 
@@ -29,14 +29,20 @@ module State =
 
 module App =
     open State
+    
+    let isValidEmail str = 
+        let emailRegex = @"\A(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?)\Z"
+        System.Text.RegularExpressions.Regex.IsMatch(str, emailRegex, System.Text.RegularExpressions.RegexOptions.IgnoreCase)
+
+    let isInt str = Int32.TryParse str |> fst
 
     let view _ _ =
-        [ "Name" |> Binding.twoWay (fun m -> m.Name) (fun v m -> SetName v)
-          "Email" |> Binding.twoWay (fun m -> m.Email) (fun v m -> SetEmail v)
-          "Age" |> Binding.twoWay (fun m -> string m.Age) (fun v m -> SetAge v)
+        [ "Name" |> Binding.twoWayValidation (fun m -> m.Name) (fun v m -> if v.Length > 2 then SetName v |> Ok else Error "Name is invalid")
+          "Email" |> Binding.twoWayValidation (fun m -> m.Email) (fun v m -> if isValidEmail v then SetEmail v |> Ok else Error "Email is invalid")
+          "Age" |> Binding.twoWayValidation (fun m -> string m.Age) (fun v m -> if isInt v then SetAge v |> Ok else Error "Age is invalid")
           "Address" |> Binding.twoWay (fun m -> m.Address) (fun v m -> SetAddress v)
           "Phone" |> Binding.twoWay (fun m -> m.Phone) (fun v m -> SetPhone v)
-          "Postal" |> Binding.twoWay (fun m -> m.Postal) (fun v m -> SetPostal v)
+          "Postal" |> Binding.twoWayValidation (fun m -> m.Postal) (fun v m -> if v.Length > 6 then Error "Postal code is invalid" else SetPostal v |> Ok)
 
           "IsEditable" |> Binding.oneWay (fun m -> not m.IsDialogVisible)
           "IsDialogVisible" |> Binding.oneWay (fun m -> m.IsDialogVisible)
@@ -50,6 +56,7 @@ module App =
     [<EntryPoint;STAThread>]
     let main argv = 
         Program.mkSimple init update view
+        |> Program.withMessageBoxErrorHandler
 //        |> Program.withConsoleTrace
 //        |> Program.withSubscription subscribe
         |> Program.runWindow (MainWindow())